requirements.txt
icalendar>=6.0.0
httpx>=0.27.0
apscheduler>=3.10.0
pyyaml>=6.0
pydantic>=2.0.0
